Update persistent damage editor to app v2
