+++ src/styles/item/_index.scss
@@ -580,4 +580,4 @@
     }
 }
 
-@import "item-attacher", "persistent-damage-dialog";
+@import "item-attacher", "persistent-damage-editor";
