+++ static/lang/en.json
@@ -1863,6 +1863,7 @@
                     },
                     "Dialog": {
                         "Add": "Add Persistent Damage",
+                        "Edit": "Edit Persistent Damage",
                         "Invalid": "Formula is invalid.",
                         "Remove": "Remove Persistent Damage",
                         "Title": "Persistent Damage ({actor})"
